Tidy up config
