+++ src/components/Blocks/Card/index.js
@@ -1,3 +1,3 @@
 export { CardEdit } from './Edit';
 export { CardView } from './View';
-export { cardSchema } from './schema';
+export * from './schema';
